app/requirements.txt
Flask-RESTful==0.3.9
Werkzeug==2.2.2
Flask-SQLAlchemy==3.0.2
Flask-Migrate==4.0.0
flask-marshmallow==0.14.0
Flask-Authorize==0.2.6
Flask-Cors==3.0.10
marshmallow-sqlalchemy==0.28.1
SQLAlchemy==2.0.22
structlog==22.2.0
pymysql==1.1.0